refactor: move validators to governance/object.h; remove lonely empty header and cpp
