+++ src/test/governance_validators_tests.cpp
@@ -2,7 +2,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include <governance/validators.h>
+#include <governance/object.h>
 #include <util/strencodings.h>
 
 #include <test/data/proposals_invalid.json.h>
@@ -46,16 +46,18 @@ BOOST_AUTO_TEST_CASE(valid_proposals_test)
         // legacy format
         std::string strHexData1 = CreateEncodedProposalObject(objProposal);
         std::string strErrorMessages;
-        BOOST_CHECK(!ValidateProposal(strHexData1, strErrorMessages, /*fCheckExpiration=*/true, fAllowScript));
+        BOOST_CHECK(!governance::ValidateProposal(strHexData1, strErrorMessages, /*fCheckExpiration=*/true, fAllowScript));
         BOOST_CHECK_EQUAL(strErrorMessages, "Proposal must be a JSON object;JSON parsing error;");
 
         // current format
         std::string strHexData2 = HexStr(objProposal.write());
         std::string strErrorMessages2;
-        BOOST_CHECK_MESSAGE(ValidateProposal(strHexData2, strErrorMessages2, /*fCheckExpiration=*/false, fAllowScript),
+        BOOST_CHECK_MESSAGE(governance::ValidateProposal(strHexData2, strErrorMessages2, /*fCheckExpiration=*/false,
+                                                         fAllowScript),
                             strErrorMessages2);
         std::string strErrorMessages3;
-        BOOST_CHECK_MESSAGE(!ValidateProposal(strHexData2, strErrorMessages3, /*fCheckExpiration=*/true, fAllowScript),
+        BOOST_CHECK_MESSAGE(!governance::ValidateProposal(strHexData2, strErrorMessages3, /*fCheckExpiration=*/true,
+                                                          fAllowScript),
                             strErrorMessages3);
     }
 }
@@ -73,17 +75,18 @@ BOOST_AUTO_TEST_CASE(invalid_proposals_test)
         // legacy format
         std::string strHexData1 = CreateEncodedProposalObject(objProposal);
         std::string strErrorMessages1;
-        BOOST_CHECK(!ValidateProposal(strHexData1, strErrorMessages1, /*fCheckExpiration=*/true, /*fAllowScript=*/false));
+        BOOST_CHECK(!governance::ValidateProposal(strHexData1, strErrorMessages1, /*fCheckExpiration=*/true,
+                                                  /*fAllowScript=*/false));
         std::string strErrorMessages1b;
-        BOOST_CHECK_MESSAGE(!ValidateProposal(strHexData1, strErrorMessages1b, /*fCheckExpiration=*/false,
-                                              /*fAllowScript=*/false),
+        BOOST_CHECK_MESSAGE(!governance::ValidateProposal(strHexData1, strErrorMessages1b, /*fCheckExpiration=*/false,
+                                                          /*fAllowScript=*/false),
                             strErrorMessages1b);
 
         // current format
         std::string strHexData2 = HexStr(objProposal.write());
         std::string strErrorMessages2;
-        BOOST_CHECK_MESSAGE(!ValidateProposal(strHexData2, strErrorMessages2, /*fCheckExpiration=*/false,
-                                              /*fAllowScript=*/false),
+        BOOST_CHECK_MESSAGE(!governance::ValidateProposal(strHexData2, strErrorMessages2, /*fCheckExpiration=*/false,
+                                                          /*fAllowScript=*/false),
                             strErrorMessages2);
     }
 }
